refactor : MoveStrategy 패키지를 domain으로 변경
