+++ src/main/java/domain/moveStrategy/NumberMoveStrategy.java
@@ -1,4 +1,4 @@
-package domain.car.moveStrategy;
+package domain.moveStrategy;
 
 import domain.numbergenerator.NumberGenerator;
